Add files via upload
My first working copy of the program that will read, clean, and format the data in the Excel files. At the moment, the final dataframe produced by this program has 2 copies of each game in it. I may choose to keep it this way to save me having to read multiple columns for team names or I may modify this at a later date to cut the output data in half.