Added Regression.py
This file marks the beginning of the machine learning content in this project. This file simply constructs a preliminary multilinear regression model. Expect this model to be improved and updated as I gain more information and get more comfortable in this project. I will also be working to collect test data to get a rough testing accuracy for this model.