OilersPanthersDistributions.R Uploaded
This R script generates distributions for each statistic required for Regression.py to generate projections. After distributions are generated, 10000 samples are generated for each statistic. This work naively assumes that all statistics are independent. This is not a good assumption to make, so this serves more as a preliminary rough model rather than a fully serious model.